refactor: move frontend actions to tool folder
